Leaving It in This State
I'm probably just going to leave this project as is. It's not very functional, but I don't feel like continuing it.

To give a note about what isn't working, the colliders interfere with one another, so the behaviours and physical collisions don't work properly. And being honest the behaviour programming is messy anyway.

Not sure if I'll ever come back to this, but I'm leaving it as is.